Initial commit to move to cmake
This commit:
* adds cmakelists files
* updates locations of some header files in #include-s
* updates genkinds.pl to make it a bit cleaner and allow an extra
  option, needed for this compilation
* adds automatic version number from git (release will have to do it
  differently, build should not fail for lack of git)

But it also breaks the make-based build (do we want it? If so, all we
need to fix are the #include-s). Also it lacks tests (CTest should be
used). Also, it doesn't check for bison and flex.

The build works with -j100, parallel builds should now succeed
without a problem (dependencies are correctly handled)