CMake modules for identifying hash_set and hash_map
Part of the work needed for #21. The modules define macros for looking for
hash_set and hash_map under a variety of headers, namespaces, and class names.

The modules have only been tested on Clang 3.3svn on OS X, but are hopefully
easy to tweak if I got something wrong.

This also introduces an include/config.h file generated by CMake. (There may be
a better place to put it.) The remaining work involves changing source files to
use the macros defined in config.h.