Source changes to use the correct hash_set et al.
This is most of the remaining source changes for #21, but while investigating
further, I arrived at the conclusion that many of the hash functions for C
strings are incorrect, leading to #26. But, that can be addressed independently
of the CMake conversion.

The build succeeds now on Clang 3.3svn on OS X, and needs to be tested on at
least Linux as well.

The CheckCxxHash*.cmake modules could be rewritten to remove redundancy, and
need to set the -std=c++0x flag on C++ compiler invocations, but this isn't
possible without further source changes.