define __attribute__() away...
this is gcc specific and doesn't work with msvc. Its used like
- __attribute__((noreturn)), which is safe to ignore, and
- __attribute_-((mode(TI)), which messes with the bitwidths of the types.

The second one is suspicious and might not be safe to ignore. No idea if
this is necessary on windows, too, or if this is safe to ignore. Let's just
do it and keep it in mind if STP on windows behaves strangly.