Removed the disgusting MERGE_STATIC_LIBRARIES and all uses of it.
Now when libstp is built the boost-library will not be merged in to
it. CMake clients of libstp

* the stp executable
* Any CMake project that imports our targets

will automatically know which Boost libraries to link in.

Clients of libstp (e.g. KLEE )who do not use CMake will now have to link
in the right Boost libraries manually to their application