Added basic infrastructure to test STP's C API using GoogleTest as the driving library. These tests are run using CTest.
Because of problems with cmake the ``test`` (i.e. $ make test) target
should not be used because it does not rebuild tests. Instead the
``run_ctest`` target should be used to build tests and then run CTest.

The ``check`` target has been setup to run the query-file-tests and
the run_ctest tests.