Fix linking issues when building libstp as a shared library by removing commenting out (eurgh...) dead code.
This is based on the work of Jurriaan Bremer

The abc code calls external functions

Abc_FrameGetGlobalFrame Abc_ShowFile Fra_FraigChoice Cmd_CommandExecute
Kit_SopFactor

which are not defined anywhere which causes problems when trying to link
By removing all functions that call these external functions, and then
removing functions that call the functions we just removed and so on...

we prevent compilation of dead code that causes linking errors.

We should really just remove this library altogether and try and use the
upstream abc library.