make BUILD_SHARED_LIBS work with WIN32
previously, BUILD_SHARED_LIBS is set to ON,
then used repeatedly and then set to OFF if on
WIN32. This leads to an inconsistent state in
which, eventually, STP_STATIC_LIBRARY is set to
libstp_static and CMake tries to export this
target even though the target is not build.