replace <msvc/stdint.h> by "msvc/stdint.h"
this occurs in various headers in the cryptominisat2 directory. msvc/ is
a folder local to cryptominisat. This implementation of stdint.h is
required and provided by cryptominisat because MSVC doesn't provide this
header file at all.

Note that there is another implementaion of stdin.h in STP's windows/
directory. I have no idea what happens if these differ.