Only '-static' is needed and if not provided it's the best we can do
Only the "LINK_SEARCH_START_STATIC" was needed, and there is no need to
unset it, as it's only for one target. Furthermore the setting only
affects what cmake things of the default behaviour of the linker, and
cmake doesn't save its old assumption so it cannot be restored anyway.

The "-static" linker flag is still needed though, otherwise we cannot
get gcc_s and other libraries to link statically:

http://stackoverflow.com/questions/3762057/cmake-how-to-produce-binaries-as-static-as-possible