Large reorganisation. Header files have been split out into their own include/stp folder and library code has been moved into to the relevant folders in lib/ . The stp tool has been moved in tools/stp.
This layout mimics that of the LLVM project which is a very well
organised project.

The motivation for doing this was the disgusting relative #includes
that were all over the place. These made compiler errors/warnings hard
to read and made it impossible for an application external to STP use
the header files!

``include/stp`` is supposed to contain public header files that would be
installed so that STP can really be used as a library. This probably
isn't done correctly right now because

- Some headers that should be internal should be in lib/
- The use of the extlib-abc, extlib-constbv mess this up. We should
  really be using these as a proper external library that the user
  has installed and build against those header files. We also shouldn't
  be bundling this library inside STP!
- The minisat and cryptominisat2 SAT solvers are mostly in lib/.
  This feels messy, again we should really building against installed
  versions of this library and not bundle the library inside STP!