Add copyright notices to source files missing them. This was reported by Matthew Dempsky.
I'd like to redo all the copyright headers at some point because the
"AUTHORS" thing is not useful (that's what git blame is for!). I'd also
like a textual description of what the code in each file is for/does.

I think following the LLVM copyright header text style would be a good
idea.