Clang-format the code base. STP's code was a big mess of diferent styles. This unifies the style based on our .clang-format style file.
This might make using ``git blame`` a little harder. You can use the
``-w`` to ignore whitespace changes which works for many lines that
have been changed by this commit but not all. For lines where ``-w``
does not work you need tell git blame to look at commits before this
one. For example

$ git blame XXXXXX^ include/stp/AST/AST.h

where XXXXXX is the commit hash of this commit.