removed never-interactive in smt2.lex ...
Without never-interactive, lex generates code that calls the function
isatty(), which fails with Visual Studio 2013 (didn't check for other
version). According to [1], isatty() is deprecated and _isatty() should be
used instead. _isatty() is in io.h, so this patch inludes io.h and defines
isatty() to call _isatty() for windows.

Seems to work fine on my system now, but did not check on Linux. This
patch only does this for smt2, not for smt or cvc, because AFAIK those are not
interactive.

[1] http://msdn.microsoft.com/library/ms235388.aspx