Cleaning up crypto examples
The T3..T5 are not clear at all what they represent -- in any case, they
do not seem interesting at all and could be anything (crypto or
otherwise)

The examples that are too slow should not be in our test set. We want
these to be relevant and useable, not stale and overly complicated to
execute.

The tea.pl for the TEA cipher generation is great and the stuff it
generates looks good and readable. I suggest we keep it.

The RSA examples I am not familiar with. They look complicated, but I
need more time. For the moment, let's keep them inside. A generator
(like tea.pl) would be way better than the generated problems
themselves.

Related to issuse #82