Now I understand fast_exit
It checks whether we can make _either_ of the two values bigger or
smaller than the other. So it only returns true if it finds the 2
inputs to be:

XXXXXX*
XXXXXX*

where two X's in the same column have to match. E.g.:

111010*
111010*

return true. ANY other stuff return false. Now, the edge case is where
we have:

1110100
1110100

I.e. they are all the same but no '*' ever comes. In this case, the
system cannot make either bigger or smaller on a whim, so we must return
FALSE.

By the way, the if-then-else-else-else... was really confusing and
that's why this thing was hiding so long.

Fixes #47