Fixing issue with l_Undef in older MiniSat
This fixes issues with older MiniSat from Ubuntu 14.04 repositories
misbehaving and not having proper namespace for the
l_Undef/l_True/l_False plus not having printStats(). We now cannot print
stats from MiniSat but at least most compilation problems should be
fixed and the solver now works out-of-the-box for a lot of default
installs.

Closes issue #172