Fix vc_getCounterExample for arbitrary expressions
Currently vc_getCounterExample only works for constant expressions,
boolean expressions, and for bitvector expressions that happen to be
in the counterexample map. If none of these cases hold, it simply (and
confusingly) returns a constant 0 expression. However, the prototype
of vc_getCounterExample suggests that it can compute an arbitrary
expression's value in the counterexample and the documentation does
nothing to dissuade this. Furthermore, a caller generally doesn't know
what's in the counterexample map, which limits it to usefully querying
only constant and variable expressions and nothing more complex.

All of the mechanism to get the counterexample value of an arbitrary
expression already exists. In fact, by substantially *simplifying*
AbsRefine_CounterExample::GetCounterExample to just call
ComputeFormulaUsingModel or TermToConstTermUsingModel according to the
expression type, we can make it far more general. Most of the logic
currently in GetCounterExample either duplicates logic in
TermToConstTermUsingModel or needlessly restricts it, so by simply
ripping it out, we make GetCounterExample work on arbitrary
expressions.

We also add a test that checks the computed values of various
expressions in the counterexample. Without the change to
GetCounterExample, this test fails.

Fixes #203.