SAT: allow to build with Riss
Instead of Minisat or CryptoMinisat, Riss could be used as backend. A build
command could look as follows, where the variable RISS_DIR should point to
the base directory of Riss:

mkdir -p build
cd build
cmake .. -DUSE_RISS=ON -DNOCRYPTOMINISAT=ON -DRISS_DIR=~/git/riss/

For the dynamic build, shared object files are required, that have to be
compiled with -fpic. As Riss does not offer a single .so file, split the
cases.