starexec: add portfolio script
When running with multiple cores, a plain portfolio approach for
several SAT solvers is a viable approach for a small amount of
solvers. When more resources are available, smarter solutions with
sharing should be considered.

This commit adds a script that implements the above idea to a certain
extend. Based on the available resources, the SAT backend uses
the solvers CryptoMiniSat, Riss and MiniSat in parallel. While Riss and
MiniSat get at most one core, CryptoMiniSat is assigned all remainint
resources. This way, we use several solvers, and CryptoMiniSat can share
information about its parallel solver threads.