`ToSATAIG::runSolver` now returns the value of `SATSolver::solve`
Issue #396 documented a case where STP was crashing when using a build
of CMS that had 'BreakID' enabled. Such a configuration of CMS can
return `l_False` for a `solve` check, but *also* return `true` on a call
to `okay`.

This led to a crash in `ConstructCounterExample` where STP expected the
underlying SAT solver to have a satisfying model, when actually it had
reported UNSAT.

This commit makes it such that `ToSATAIG::runSolver` results the value
of `SATSolver::solve` and not `SATSolver::okay`.

Signed-off-by: Andrew V. Jones <andrewvaughanj@gmail.com>