container: add competition script
Let's make sure we can test the solver locally via a file specified in
the environment, as well as via a file in S3. Hence, add a script that
can handle both cases and next calls the portfolio solver that is also
used in starexec.

While this case also covers the stdin case, the tooling is currently
broken.

Signed-off-by: Norbert Manthey <nmanthey@conp-solutions.com>