CMakeLists: use absolute libdir in rpath handling
Commit a26083b634fe (Fixing libdirs for CMakeLists) switched from
"${CMAKE_INSTALL_PREFIX}/lib" to "${CMAKE_INSTALL_LIBDIR}". It fixed the
issue with lib vs lib64. But while the former is absolute, the latter
needs not. That causes troubles as rpath can be set to something like
"lib64" only. That is bogus.

So fix this by using absolute "${CMAKE_INSTALL_FULL_LIBDIR}" from
GNUInstallDirs which is already included and handles the paths
correctly.

Fixes #437.