Make timestamp in generated file use $SOURCE_DATE_EPOCH
Currently the generator puts the time of the build in the file, making
the results different in each rebuild. Use $SOURCE_DATE_EPOCH to set
the timestamp, if the variable is defined.

See https://reproducible-builds.org/docs/source-date-epoch/ for
background and the definition of the variable.

The code modification is based on
https://reproducible-builds.org/docs/source-date-epoch/#perl.

Also, fix typo in the header.